+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All Notable changes to `jobs-careercast` will be documented in this file
 ## 1.0.1 - 2016-10-06
 
 ### Fixed
-- DatePosted on Job object now being saved as object instead of string. 
+- DatePosted and ValidThrough on Job object now being saved as object instead of string. 
 
 ## 1.0.0 - 2016-09-22